+++ src/components/layout/Header.js
@@ -1,6 +1,5 @@
-import { AppBar, Typography, Toolbar, Button, MenuItem, Menu } from "@mui/material";
+import { AppBar, Typography, Toolbar } from "@mui/material";
 import { makeStyles } from "@mui/styles";
-import logoSD from '../../resources/logoIdea.PNG';
 import { useNavigate } from "react-router";
 import { COLOR_PALETTE } from "../../Constants";
 import React, { useState } from "react";
@@ -61,24 +60,24 @@ const Header = () => {
         navigate('/create-account');
     }
 
-    // TODO
-    async function handleTemp () {
-        let res = await saveUser({
-            fname: 'test',
-            lname: 'icle',
-            email: 'testicle@domain.cum'+Math.floor(Math.random()*100),
-            password: 'temp123'
-        })
-        console.log(res);
-    }
+    // // TODO
+    // async function handleTemp () {
+    //     let res = await saveUser({
+    //         fname: 'test',
+    //         lname: 'icle',
+    //         email: 'testicle@domain.cum'+Math.floor(Math.random()*100),
+    //         password: 'temp123'
+    //     })
+    //     console.log(res);
+    // }
 
-    async function handleTemp2 () {
-        let res = await login({
-            email: 'testicle@domain.cum',
-            password: 'temp123'
-        })
-        console.log(res);
-    }
+    // async function handleTemp2 () {
+    //     let res = await login({
+    //         email: 'testicle@domain.cum',
+    //         password: 'temp123'
+    //     })
+    //     console.log(res);
+    // }
 
     const handleHome = () => {
         navigate('/');