+++ src/CHANGELOG.md
@@ -6,7 +6,8 @@ Create a new version whenever we update main/master.
 Add items closer to the version number so the date go from newest to oldest (top to bottom).
 
 ### *Version 0.1.1*
- - 2022/01/09 - [Header Changes to support before and after login](https://sd-onlinepos.atlassian.net/browse/ON-75) 
+ - 2023/01/10 - [Create Layout for Create Business Page](https://sd-onlinepos.atlassian.net/browse/ON-73) 
+ - 2023/01/09 - [Header Changes to support before and after login](https://sd-onlinepos.atlassian.net/browse/ON-75) 
  - 2022/11/28 - [Create Reuseable components](https://sd-onlinepos.atlassian.net/browse/ON-28) 
  - 2022/11/28 - [Create Layout for Landing Page](https://sd-onlinepos.atlassian.net/browse/ON-41) 
  - 2022/11/20 - [Create Layout for Create Account Page](https://sd-onlinepos.atlassian.net/browse/ON-45)